Extract KeyValue and ChannelValueType into a dedicated KeyValue.h header
